+++ server/src-lib/Hasura/RQL/DML/Returning.hs
@@ -93,12 +93,8 @@ mkDefaultMutFlds = MOutMultirowFields . \case
   where
     mutFlds = [("affected_rows", MCount)]
 
-qualTableToAliasIden :: QualifiedTable -> Iden
-qualTableToAliasIden qt =
-  Iden $ snakeCaseTable qt <> "__mutation_result_alias"
-
-mkMutFldExp :: QualifiedTable -> Maybe Int -> Bool -> MutFld -> S.SQLExp
-mkMutFldExp qt preCalAffRows strfyNum = \case
+mkMutFldExp :: Iden -> Maybe Int -> Bool -> MutFld -> S.SQLExp
+mkMutFldExp cteAlias preCalAffRows strfyNum = \case
   MCount ->
     let countExp = S.SESelect $
           S.mkSelect
@@ -112,28 +108,70 @@ mkMutFldExp qt preCalAffRows strfyNum = \case
         tabPerm = TablePerm annBoolExpTrue Nothing
     in S.SESelect $ mkSQLSelect JASMultipleRows $
        AnnSelG selFlds tabFrom tabPerm noTableArgs strfyNum
-  where
-    cteAlias = qualTableToAliasIden qt
 
+{- Note [Mutation output expression]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+An example output expression for INSERT mutation:
+
+WITH "<table-name>__mutation_result_alias" AS (
+  INSERT INTO <table-name> (<insert-column>[..])
+  VALUES
+    (<insert-value-row>[..])
+    ON CONFLICT ON CONSTRAINT "<table-constraint-name>" DO NOTHING RETURNING *,
+    -- An extra column expression which performs the 'CHECK' validation
+    CASE
+      WHEN (<CHECK Condition>) THEN NULL
+      ELSE "hdb_catalog"."check_violation"('insert check constraint failed')
+    END
+),
+"<table-name>__all_columns_alias" AS (
+  -- Only extract columns from mutated rows. Columns sorted by ordinal position so that
+  -- resulted rows can be casted to table type.
+  SELECT (<table-column>[..])
+  FROM
+    "<table-name>__mutation_result_alias"
+)
+<SELECT statement to generate mutation response using '<table-name>__all_columns_alias' as FROM>
+-}
+
+-- | Generate mutation output expression with given mutation CTE statement.
+-- See Note [Mutation output expression].
 mkMutationOutputExp
-  :: QualifiedTable -> Maybe Int -> S.CTE -> MutationOutput -> Bool -> S.SelectWith
-mkMutationOutputExp qt preCalAffRows cte mutOutput strfyNum =
-  S.SelectWith [(S.Alias cteAlias, cte)] sel
+  :: QualifiedTable
+  -> [PGColumnInfo]
+  -> Maybe Int
+  -> S.CTE
+  -> MutationOutput
+  -> Bool
+  -> S.SelectWith
+mkMutationOutputExp qt allCols preCalAffRows cte mutOutput strfyNum =
+  S.SelectWith [ (S.Alias mutationResultAlias, cte)
+               , (S.Alias allColumnsAlias, allColumnsSelect)
+               ] sel
   where
-    cteAlias = qualTableToAliasIden qt
+    mutationResultAlias = Iden $ snakeCaseQualObject qt <> "__mutation_result_alias"
+    allColumnsAlias = Iden $ snakeCaseQualObject qt <> "__all_columns_alias"
+
+    allColumnsSelect = S.CTESelect $ S.mkSelect
+                       { S.selExtr = map S.mkExtr $ map pgiColumn $ sortCols allCols
+                       , S.selFrom = Just $ S.mkIdenFromExp mutationResultAlias
+                       }
+
     sel = S.mkSelect { S.selExtr = [S.Extractor extrExp Nothing] }
+          where
+            extrExp = case mutOutput of
+              MOutMultirowFields mutFlds ->
+                let jsonBuildObjArgs = flip concatMap mutFlds $
+                      \(FieldName k, mutFld) -> [ S.SELit k
+                                                , mkMutFldExp allColumnsAlias preCalAffRows strfyNum mutFld
+                                                ]
+                in S.SEFnApp "json_build_object" jsonBuildObjArgs Nothing
 
-    extrExp = case mutOutput of
-      MOutMultirowFields mutFlds ->
-        let jsonBuildObjArgs = flip concatMap mutFlds $
-              \(FieldName k, mutFld) -> [S.SELit k, mkMutFldExp qt preCalAffRows strfyNum mutFld]
-        in S.SEFnApp "json_build_object" jsonBuildObjArgs Nothing
-
-      MOutSinglerowObject annFlds ->
-        let tabFrom = FromIden cteAlias
-            tabPerm = TablePerm annBoolExpTrue Nothing
-        in S.SESelect $ mkSQLSelect JASSingleObject $
-           AnnSelG annFlds tabFrom tabPerm noTableArgs strfyNum
+              MOutSinglerowObject annFlds ->
+                let tabFrom = FromIden allColumnsAlias
+                    tabPerm = TablePerm annBoolExpTrue Nothing
+                in S.SESelect $ mkSQLSelect JASSingleObject $
+                   AnnSelG annFlds tabFrom tabPerm noTableArgs strfyNum
 
 
 checkRetCols