+++ CHANGELOG.md
@@ -104,3 +104,4 @@
 - option to reload remote schemas in 'reload_metadata' API (fix #3792, #4117)
 - console: disable selecting roles without permissions for bulk actions (close #4178) (#4195)
 - console: show remote schema / event trigger intro sections always (#4044)
+- server: fix postgres query error when computed fields included in mutation response (fix #4035)